+++ Tracker/HabitCreating/TextFieldCell.swift
@@ -73,4 +73,9 @@ extension TextFieldCell: UITextFieldDelegate {
             delegate?.didTextChange(text: textField.text)
         }
     }
+    
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        textField.resignFirstResponder()
+        return true
+    }
 }
