--- Tracker/Store/DaysValueTransformer.swift
@@ -1,32 +1,0 @@
-//
-//  DaysValueTransformer.swift
-//  Tracker
-//
-//  Created by ulyana on 21.08.25.
-//
-
-import Foundation
-
-@objc
-final class DaysValueTransformer: ValueTransformer {
-    
-    override class func transformedValueClass() -> AnyClass { NSData.self }
-    override class func allowsReverseTransformation() -> Bool { true }
-    
-    override func transformedValue(_ value: Any?) -> Any? {
-        guard let days = value as? [Int] else { return nil }
-        return try? JSONEncoder().encode(days)
-    }
-
-    override func reverseTransformedValue(_ value: Any?) -> Any? {
-        guard let data = value as? NSData else { return nil }
-        return try? JSONDecoder().decode([Int].self, from: data as Data)
-    }
-    
-    static func register() {
-        ValueTransformer.setValueTransformer(
-            DaysValueTransformer(),
-            forName: NSValueTransformerName(rawValue: String(describing: DaysValueTransformer.self))
-        )
-    }
-}
