+++ Tracker/Store/Models/TrackerCategoryStoreUpdate .swift
@@ -1,0 +1,13 @@
+//
+//  TrackerCategoryStoreUpdate .swift
+//  Tracker
+//
+//  Created by ulyana on 18.09.25.
+//
+
+import Foundation
+
+struct TrackerCategoryStoreUpdate {
+    let insertedIndexes: IndexSet
+    let deletedIndexes: IndexSet
+}
