+++ Tracker/Store/Models/TrackerRecordStoreUpdate.swift
@@ -1,0 +1,13 @@
+//
+//  TrackerRecordStoreUpdate.swift
+//  Tracker
+//
+//  Created by ulyana on 18.09.25.
+//
+
+import Foundation
+
+struct TrackerRecordStoreUpdate {
+    let insertedIndexes: IndexSet
+    let deletedIndexes: IndexSet
+}
