+++ Tracker/Store/Models/TrackerStoreUpdate.swift
@@ -1,0 +1,13 @@
+//
+//  TrackerStoreUpdate.swift
+//  Tracker
+//
+//  Created by ulyana on 18.09.25.
+//
+
+import Foundation
+
+struct TrackerStoreUpdate {
+    let insertedIndexes: IndexSet
+    let deletedIndexes: IndexSet
+}
