+++ Tracker/Store/TrackerCategoryStoreError.swift
@@ -1,0 +1,12 @@
+//
+//  TrackerCategoryStoreError.swift
+//  Tracker
+//
+//  Created by ulyana on 18.09.25.
+//
+
+import Foundation
+
+enum TrackerCategoryStoreError: Error {
+    case decodingErrorInvalidTitle
+}
