+++ Tracker/Store/TrackerRecordStoreError.swift
@@ -1,0 +1,14 @@
+//
+//  TrackerRecordStoreError.swift
+//  Tracker
+//
+//  Created by ulyana on 18.09.25.
+//
+
+import Foundation
+
+enum TrackerRecordStoreError: Error {
+    case decodingErrorInvalidId
+    case decodingErrorInvalidDate
+    case deleteRecordError
+}
